requirements.txt
# MA分割模型训练所需依赖包

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# YOLO框架
ultralytics>=8.0.0

# 图像处理
opencv-python>=4.7.0
Pillow>=9.5.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 进度条和可视化
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 其他工具
pyyaml>=6.0
scikit-learn>=1.2.0

# 可选：加速和优化
# albumentations>=1.3.0  # 高级数据增强
# tensorboard>=2.13.0     # 训练可视化

